fix: organized status components on the one folder
